feat: added api keys

+++ src/project/project.controller.ts
@@ -15,6 +15,7 @@ import {
   ApiBadRequestResponse,
   ApiBearerAuth,
   ApiCreatedResponse,
+  ApiHeader,
   ApiOkResponse,
   ApiTags,
 } from '@nestjs/swagger';
@@ -34,31 +35,12 @@ import { Role } from 'auth/decorators/roles.enum';
 import { GetProjectsDTO } from './dtos/get-projects';
 import { AddProjectMemberDTO } from './dtos/add-project-member';
 import { RemoveProjectMemberDTO } from './dtos/remove-project-member';
+import { Messages } from 'utils/constants';
 
 @Controller('projects')
 @ApiTags('projects')
 export class ProjectController {
   constructor(private projectService: ProjectService) {}
-
-  @Get()
-  @ApiBearerAuth('defaultBearerAuth')
-  @UseGuards(AuthGuard, JwtGuard)
-  @Roles(Role.SuperAdmin, Role.Member, Role.Admin)
-  @ApiOkResponse({
-    description: constants.OK.description,
-    type: GetProjectsDTO,
-  })
-  @ApiBadRequestResponse({
-    description: constants.BAD_REQUEST.description,
-    type: BadRequestDTO,
-  })
-  async getAllProjects() {
-    try {
-      return { projects: await this.projectService.getAllProjects() };
-    } catch (err) {
-      throw new BadRequestException(err.message);
-    }
-  }
 
   @Post('create')
   @ApiBearerAuth('defaultBearerAuth')
@@ -92,6 +74,7 @@ export class ProjectController {
   }
 
   @Post('remove')
+  @ApiHeader({ name: 'app_id', example: '' })
   @ApiBearerAuth('defaultBearerAuth')
   @UseGuards(AuthGuard, JwtGuard)
   @Roles(Role.SuperAdmin, Role.Admin)
@@ -109,6 +92,18 @@ export class ProjectController {
     @Req() req,
   ) {
     const { projectName } = projectDetails;
+    if (!req.headers?.app_id) {
+      throw new BadRequestException(Messages.AppIdRequired);
+    }
+
+    if (
+      !(await this.projectService.validateAppId(
+        req.headers.app_id,
+        projectName,
+      ))
+    ) {
+      throw new Error(Messages.IncorrectAppId);
+    }
 
     try {
       return this.projectService.removeProject(req?.user?.id, projectName);
@@ -118,6 +113,7 @@ export class ProjectController {
   }
 
   @Get('info/:projectName')
+  @ApiHeader({ name: 'app_id', example: '' })
   @ApiBearerAuth('defaultBearerAuth')
   @UseGuards(AuthGuard, JwtGuard)
   @Roles(Role.SuperAdmin, Role.Member, Role.Admin)
@@ -137,6 +133,19 @@ export class ProjectController {
     const { projectName } = projectDetails;
 
     try {
+      if (!req.headers?.app_id) {
+        throw new BadRequestException(Messages.AppIdRequired);
+      }
+
+      if (
+        !(await this.projectService.validateAppId(
+          req.headers.app_id,
+          projectName,
+        ))
+      ) {
+        throw new Error(Messages.IncorrectAppId);
+      }
+
       return this.projectService.getProjectDetails(req?.user.id, projectName);
     } catch (err) {
       throw new BadRequestException(err.message);
@@ -144,6 +153,7 @@ export class ProjectController {
   }
 
   @Post('/events/add')
+  @ApiHeader({ name: 'app_id', example: '' })
   @ApiBearerAuth('defaultBearerAuth')
   @UseGuards(AuthGuard, JwtGuard)
   @Roles(Role.SuperAdmin, Role.Member, Role.Admin)
@@ -162,6 +172,21 @@ export class ProjectController {
         webhook_url,
         sync_historical_data = false,
       } = addEventDetails;
+
+      if (!req.headers?.app_id) {
+        throw new BadRequestException(Messages.AppIdRequired);
+      }
+
+      if (
+        !(await this.projectService.validateAppId(
+          req.headers.app_id,
+          null,
+          projectId,
+        ))
+      ) {
+        throw new Error(Messages.IncorrectAppId);
+      }
+
       await this.projectService.addEvent(
         req?.user.id,
         projectId,
@@ -179,6 +204,7 @@ export class ProjectController {
   }
 
   @Post('/events/remove')
+  @ApiHeader({ name: 'app_id', example: '' })
   @ApiBearerAuth('defaultBearerAuth')
   @UseGuards(AuthGuard, JwtGuard)
   @Roles(Role.SuperAdmin, Role.Member, Role.Admin)
@@ -192,6 +218,21 @@ export class ProjectController {
   ) {
     try {
       const { name, projectId } = removeEventDetails;
+
+      if (!req.headers?.app_id) {
+        throw new BadRequestException(Messages.AppIdRequired);
+      }
+
+      if (
+        !(await this.projectService.validateAppId(
+          req.headers.app_id,
+          null,
+          projectId,
+        ))
+      ) {
+        throw new Error(Messages.IncorrectAppId);
+      }
+
       await this.projectService.removeEvent(req?.user.id, projectId, name);
     } catch (err) {
       throw new BadRequestException(err.message);
@@ -199,6 +240,7 @@ export class ProjectController {
   }
 
   @Post('/members/add')
+  @ApiHeader({ name: 'app_id', example: '' })
   @ApiBearerAuth('defaultBearerAuth')
   @UseGuards(AuthGuard, JwtGuard)
   @Roles(Role.SuperAdmin, Role.Admin)
@@ -211,6 +253,21 @@ export class ProjectController {
   ) {
     try {
       const { projectId, memberId } = addProjectMemberDto;
+
+      if (!req.headers?.app_id) {
+        throw new BadRequestException(Messages.AppIdRequired);
+      }
+
+      if (
+        !(await this.projectService.validateAppId(
+          req.headers.app_id,
+          null,
+          projectId,
+        ))
+      ) {
+        throw new Error(Messages.IncorrectAppId);
+      }
+
       await this.projectService.addMember(projectId, req?.user.id, memberId);
     } catch (err) {
       throw new BadRequestException(err.message);
@@ -218,6 +275,7 @@ export class ProjectController {
   }
 
   @Post('/members/remove')
+  @ApiHeader({ name: 'app_id', example: '' })
   @ApiBearerAuth('defaultBearerAuth')
   @UseGuards(AuthGuard, JwtGuard)
   @Roles(Role.SuperAdmin, Role.Admin)
@@ -230,6 +288,21 @@ export class ProjectController {
   ) {
     try {
       const { projectId, memberId } = removeMemberDto;
+
+      if (!req.headers?.app_id) {
+        throw new BadRequestException(Messages.AppIdRequired);
+      }
+
+      if (
+        !(await this.projectService.validateAppId(
+          req.headers.app_id,
+          null,
+          projectId,
+        ))
+      ) {
+        throw new Error(Messages.IncorrectAppId);
+      }
+
       await this.projectService.removeMember(projectId, req?.user.id, memberId);
     } catch (err) {
       throw new BadRequestException(err.message);
